requirements.txt
# ETL Pipeline Dependencies

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
python-dotenv>=0.19.0

# Database dependencies
psycopg2-binary>=2.9.0  # PostgreSQL support

# Development dependencies
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.5b2
flake8>=3.9.2
isort>=5.9.1
